# perf(cast): improve vanity wallet generation