--- a/cli/src/cmd/cast/wallet/vanity.rs
+++ b/cli/src/cmd/cast/wallet/vanity.rs
@@ -8,11 +8,7 @@ use ethers::{
     types::{H160, U256},
     utils::{get_contract_address, secret_key_to_address},
 };
-
-use rayon::{
-    iter::{ParallelBridge, ParallelIterator},
-    prelude::IntoParallelIterator,
-};
+use rayon::{iter::ParallelIterator, prelude::IntoParallelIterator};
 use regex::Regex;
 use std::{
     sync::atomic::{AtomicU64, Ordering},
@@ -136,13 +132,7 @@ impl Cmd for VanityArgs {
 }
 
 fn find_vanity_address<T: VanityMatcher>(matcher: T) -> Option<LocalWallet> {
-    repeater()
-        .find_any(|(_, addr)| {
-            let m = matcher.is_match(addr);
-            check();
-            m
-        })
-        .map(|(key, _)| key.into())
+    repeater().find_any(|(_, addr)| matcher.is_match(addr)).map(|(key, _)| key.into())
 }
 
 fn find_vanity_address_with_nonce<T: VanityMatcher>(matcher: T, nonce: u64) -> Option<LocalWallet> {
@@ -150,36 +140,36 @@ fn find_vanity_address_with_nonce<T: VanityMatcher>(matcher: T, nonce: u64) -> O
     repeater()
         .find_any(|(_, addr)| {
             let contract_addr = get_contract_address(*addr, nonce);
-            let m = matcher.is_match(&contract_addr);
-            check();
-            m
+            matcher.is_match(&contract_addr)
         })
         .map(|(key, _)| key.into())
 }
 
 fn repeater() -> impl ParallelIterator<Item = (SigningKey, H160)> {
-    // let rng = thread_rng();
-    let start = now().as_millis() as u64;
-    PREVIOUS.store(start, Ordering::Release);
-    // std::iter::repeat_with(move || {
-    //     let signer = SigningKey::random(&mut thread_rng());
-    //     let address = secret_key_to_address(&signer);
-    //     (signer, address)
-    // })
-    // .par_bridge()
-    (0..u128::MAX).into_par_iter().map(|_| {
-        let signer = SigningKey::random(&mut thread_rng());
-        let address = secret_key_to_address(&signer);
-        (signer, address)
+    setup();
+
+    (0..u128::MAX).into_par_iter().map(move |_| {
+        bench();
+
+        let key = SigningKey::random(&mut thread_rng());
+        let address = secret_key_to_address(&key);
+        (key, address)
     })
 }
 
 // TODO: remove
-const N: u64 = 10000;
+const N: u64 = 100_000;
 static PREVIOUS: AtomicU64 = AtomicU64::new(0);
 static I: AtomicU64 = AtomicU64::new(0);
 
-fn check() {
+#[inline]
+fn setup() {
+    let start = now().as_millis() as u64;
+    PREVIOUS.store(start, Ordering::Release);
+}
+
+#[inline]
+fn bench() {
     let i = I.fetch_add(1, Ordering::AcqRel);
     if i % N == 0 {
         let now = now().as_millis() as u64;